requirements.txt
astroid==2.2.5
autopep8==1.4.4
boto3==1.9.130
botocore==1.12.130
Django==1.11
django-cors-headers==3.0.1
django-filter==1.1.0
django-rest-framework==0.1.0
djangorestframework==3.9.1
djangorestframework-filters==0.11.1
docutils==0.14
isort==4.3.17
jmespath==0.9.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pkg-resources==0.0.0
psycopg2-binary==2.7.7
pycodestyle==2.5.0
pylint==2.3.1
python-dateutil==2.8.0
pytz==2018.9
s3transfer==0.2.0
six==1.12.0
typed-ast==1.3.1
urllib3==1.24.1
wrapt==1.11.1
